Rename BMI test util file extension to .hpp.
